Action bar button

+++ app/src/main/java/cry/who/boy/tap_recordatorios/Editar.java
@@ -8,6 +8,8 @@ import android.icu.util.Calendar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
@@ -18,9 +20,13 @@ import android.widget.TextView;
 import android.widget.TimePicker;
 import android.widget.Toast;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class Editar extends AppCompatActivity implements View.OnClickListener{
     private Toolbar toolbar; //Declarar el Toolbar
-    private Button btn_fecha_2, btn_hora_2, btn_cancelar_2, btn_ok_2;
+    private Button btn_fecha_2, btn_hora_2, btn_ok_2;
     private TextView tv_fecha_2, tv_hora_2;
     private EditText et_Titulo_2, et_Desc_2;
     private int dia, mes, anio, hora, minutos;
@@ -58,7 +64,6 @@ public class Editar extends AppCompatActivity implements View.OnClickListener{
         });
 
        //Botones
-        btn_cancelar_2 = (Button) findViewById(R.id.btn_cancel_2);
         btn_ok_2 = (Button) findViewById(R.id.btn_ok_2);
         btn_fecha_2 = (Button) findViewById(R.id.btn_fecha_2);
         btn_hora_2 = (Button) findViewById(R.id.btn_hora_2);
@@ -139,15 +144,16 @@ public class Editar extends AppCompatActivity implements View.OnClickListener{
             public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
                 hora=hourOfDay;
                 minutos=minute;
-                tv_hora_2.setText(hora+":"+minutos);
+                if(minutos<10){
+                    tv_hora_2.setText(hora+":0"+minutos);
+                }else{
+                    tv_hora_2.setText(hora+":"+minutos);
+                }
             }
         },hora,minutos,false);
         timePickerDialog.show();
     }
-    public void Cancelar(View view){
-        Intent intent = new Intent(Editar.this, MainActivity.class);
-        startActivity(intent);
-    }
+
     public void MandarDatos(View vista){
 
         String Titulo = et_Titulo_2.getText().toString();
@@ -155,6 +161,21 @@ public class Editar extends AppCompatActivity implements View.OnClickListener{
         String Hora = tv_hora_2.getText().toString();// <---
         String Desc = et_Desc_2.getText().toString();
         int Import =Importancia.getSelectedItemPosition();
+
+        String Fech_act = d_act+"-"+(m_act+1)+"-"+a_act;
+        java.text.DateFormat df=new SimpleDateFormat("dd-MM-yyyy");
+        java.text.DateFormat dft=new SimpleDateFormat("hh-mm");
+
+        Date startDate = null;
+        Date secondDate = null;
+
+        try{
+            startDate = df.parse(Fecha);
+            secondDate = df.parse(Fech_act);
+        }catch(ParseException e){
+            e.printStackTrace();
+        }
+
         if(Hora == null
                 || Hora.equals("")
                 || Hora.trim().length()==0){
@@ -166,20 +187,41 @@ public class Editar extends AppCompatActivity implements View.OnClickListener{
             Toast msn = Toast.makeText(getApplicationContext(), "No deje el Título o la Fecha vacío", Toast.LENGTH_SHORT);
             msn.show();
         }else {
-            Rec rec=new Rec(Titulo, Fecha, Hora, Desc, Import);
-            myDb.deleteData(tituloRec);
-            boolean veri=myDb.insertData(rec);
-            if (veri){
-                Toast msn = Toast.makeText(getApplicationContext(), "Modificado Satisfactoriamente", Toast.LENGTH_SHORT);
+            if(startDate.before(secondDate)){
+                Toast msn = Toast.makeText(getApplicationContext(), "No ponga una Fecha anterior a la del día de Hoy", Toast.LENGTH_SHORT);
                 msn.show();
-                Intent intent = new Intent(Editar.this, MainActivity.class);
-                startActivity(intent);
             }else{
-                Toast msn = Toast.makeText(getApplicationContext(), "Nel we", Toast.LENGTH_SHORT);
-                msn.show();
+                Rec rec = new Rec(Titulo, Fecha, Hora, Desc, Import);
+                myDb.deleteData(tituloRec);
+                boolean veri = myDb.insertData(rec);
+                if (veri) {
+                    Toast msn = Toast.makeText(getApplicationContext(), "Modificado Satisfactoriamente", Toast.LENGTH_SHORT);
+                    msn.show();
+                    Intent intent = new Intent(Editar.this, MainActivity.class);
+                    startActivity(intent);
+                }
             }
-
-
         }
+    }
+
+    //Este método es para saber qué opción tocó el usuario
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_borrar, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+        if(id == R.id.action_borrar){
+            myDb.deleteData(tituloRec);
+            Intent i = new Intent(Editar.this, MainActivity.class);
+            startActivity(i);
+            Toast msn = Toast.makeText(getApplicationContext(), "Eliminado Satisfactoriamente", Toast.LENGTH_SHORT);
+            msn.show();
+            return true;
+    }
+        return super.onOptionsItemSelected(item);
     }
 }
